fix the potential shift overflow problem of `machineId`

+++ cosid-core/src/main/java/me/ahoo/cosid/snowflake/ClockSyncSnowflakeId.java
@@ -97,7 +97,7 @@ public class ClockSyncSnowflakeId implements SnowflakeId {
     }
 
     @Override
-    public int getMachineId() {
+    public long getMachineId() {
         return delegate.getMachineId();
     }
