+++ cosid-core/src/main/java/me/ahoo/cosid/snowflake/MillisecondSnowflakeId.java
@@ -25,15 +25,15 @@ public class MillisecondSnowflakeId extends AbstractSnowflakeId {
     public final static int DEFAULT_MACHINE_BIT = 10;
     public final static int DEFAULT_SEQUENCE_BIT = 12;
 
-    public MillisecondSnowflakeId(int machineId) {
+    public MillisecondSnowflakeId(long machineId) {
         this(CosId.COSID_EPOCH, DEFAULT_TIMESTAMP_BIT, DEFAULT_MACHINE_BIT, DEFAULT_SEQUENCE_BIT, machineId);
     }
 
-    public MillisecondSnowflakeId(int machineBit, int machineId) {
+    public MillisecondSnowflakeId(int machineBit, long machineId) {
         super(CosId.COSID_EPOCH, DEFAULT_TIMESTAMP_BIT, machineBit, DEFAULT_SEQUENCE_BIT, machineId);
     }
 
-    public MillisecondSnowflakeId(long epoch, int timestampBit, int machineBit, int sequenceBit, int machineId) {
+    public MillisecondSnowflakeId(long epoch, int timestampBit, int machineBit, int sequenceBit, long machineId) {
         super(epoch, timestampBit, machineBit, sequenceBit, machineId);
     }
